XC-238: Create test_fixtures::arbitrary module
