+++ rs/bitcoin/ckbtc/minter/src/test_fixtures.rs
@@ -142,3 +142,137 @@ pub mod mock {
         }
     }
 }
+
+pub mod arbitrary {
+    use crate::{
+        address::BitcoinAddress, signature::EncodedSignature, state::RetrieveBtcRequest, tx,
+    };
+    use candid::Principal;
+    use ic_base_types::{CanisterId, PrincipalId};
+    use ic_btc_interface::{OutPoint, Satoshi, Txid, Utxo};
+    use icrc_ledger_types::icrc1::account::Account;
+    use proptest::{
+        array::uniform20,
+        array::uniform32,
+        collection::{vec as pvec, SizeRange},
+        option,
+        prelude::{any, Strategy},
+        prop_oneof,
+    };
+    use serde_bytes::ByteBuf;
+
+    pub fn amount() -> impl Strategy<Value = Satoshi> {
+        1..10_000_000_000u64
+    }
+
+    pub fn vec_to_txid(vec: Vec<u8>) -> Txid {
+        let bytes: [u8; 32] = vec.try_into().expect("Can't convert to [u8; 32]");
+        bytes.into()
+    }
+
+    pub fn out_point() -> impl Strategy<Value = OutPoint> {
+        (pvec(any::<u8>(), 32), any::<u32>()).prop_map(|(txid, vout)| OutPoint {
+            txid: vec_to_txid(txid),
+            vout,
+        })
+    }
+
+    pub fn unsigned_input(
+        value: impl Strategy<Value = Satoshi>,
+    ) -> impl Strategy<Value = tx::UnsignedInput> {
+        (out_point(), value, any::<u32>()).prop_map(|(previous_output, value, sequence)| {
+            tx::UnsignedInput {
+                previous_output,
+                value,
+                sequence,
+            }
+        })
+    }
+
+    pub fn signed_input() -> impl Strategy<Value = tx::SignedInput> {
+        (
+            out_point(),
+            any::<u32>(),
+            pvec(1u8..0xff, 64),
+            pvec(any::<u8>(), 32),
+        )
+            .prop_map(
+                |(previous_output, sequence, sec1, pubkey)| tx::SignedInput {
+                    previous_output,
+                    sequence,
+                    signature: EncodedSignature::from_sec1(&sec1),
+                    pubkey: ByteBuf::from(pubkey),
+                },
+            )
+    }
+
+    pub fn address() -> impl Strategy<Value = BitcoinAddress> {
+        prop_oneof![
+            uniform20(any::<u8>()).prop_map(BitcoinAddress::P2wpkhV0),
+            uniform32(any::<u8>()).prop_map(BitcoinAddress::P2wshV0),
+            uniform32(any::<u8>()).prop_map(BitcoinAddress::P2trV1),
+            uniform20(any::<u8>()).prop_map(BitcoinAddress::P2pkh),
+            uniform20(any::<u8>()).prop_map(BitcoinAddress::P2sh),
+        ]
+    }
+
+    pub fn tx_out() -> impl Strategy<Value = tx::TxOut> {
+        (amount(), address()).prop_map(|(value, address)| tx::TxOut { value, address })
+    }
+
+    pub fn utxo(amount: impl Strategy<Value = Satoshi>) -> impl Strategy<Value = Utxo> {
+        (amount, pvec(any::<u8>(), 32), 0..5u32).prop_map(|(value, txid, vout)| Utxo {
+            outpoint: OutPoint {
+                txid: vec_to_txid(txid),
+                vout,
+            },
+            value,
+            height: 0,
+        })
+    }
+
+    pub fn account() -> impl Strategy<Value = Account> {
+        (pvec(any::<u8>(), 32), option::of(uniform32(any::<u8>()))).prop_map(|(pk, subaccount)| {
+            Account {
+                owner: PrincipalId::new_self_authenticating(&pk).0,
+                subaccount,
+            }
+        })
+    }
+
+    pub fn retrieve_btc_requests(
+        amount: impl Strategy<Value = Satoshi>,
+        num: impl Into<SizeRange>,
+    ) -> impl Strategy<Value = Vec<RetrieveBtcRequest>> {
+        let request_strategy = (
+            amount,
+            address(),
+            any::<u64>(),
+            1569975147000..2069975147000u64,
+            option::of(any::<u64>()),
+            option::of(account()),
+        )
+            .prop_map(
+                |(amount, address, block_index, received_at, provider, reimbursement_account)| {
+                    RetrieveBtcRequest {
+                        amount,
+                        address,
+                        block_index,
+                        received_at,
+                        kyt_provider: provider
+                            .map(|id| Principal::from(CanisterId::from_u64(id).get())),
+                        reimbursement_account,
+                    }
+                },
+            );
+        pvec(request_strategy, num).prop_map(|mut reqs| {
+            reqs.sort_by_key(|req| req.received_at);
+
+            for (i, req) in reqs.iter_mut().enumerate() {
+                req.block_index = i as u64;
+            }
+
+            reqs
+        })
+    }
+}
